chore(color-picker): Extract sub utility from `color-picker` to `shared`
