+++ packages/clay-color-picker/package.json
@@ -28,6 +28,7 @@
 	"dependencies": {
 		"@clayui/drop-down": "^3.0.0-milestone.1",
 		"@clayui/icon": "^3.0.0-milestone.1",
+		"@clayui/shared": "^3.0.0-milestone.1",
 		"classnames": "^2.2.6",
 		"tinycolor2": "^1.4.1"
 	},
