+++ packages/clay-shared/src/index.tsx
@@ -6,7 +6,8 @@
 
 export {ClayPortal} from './Portal';
 export {getEllipsisItems} from './getEllipsisItems';
-export {useKeyHandlerForList} from './useKeyHandlerForList';
 export {useDebounce} from './useDebounce';
+export {useKeyHandlerForList} from './useKeyHandlerForList';
 export {useTransitionHeight} from './useTransitionHeight';
 export {useFocusManagement} from './useFocusManagement';
+export {sub} from './sub';
